+++ lib/gds-api-adapters.rb
@@ -1,2 +1,3 @@
 require 'gds_api/railtie' if defined?(Rails)
 require 'gds_api/exceptions'
+require 'gds_api/config'